Add files via upload
Online backup of Aristotle, a chatbot for tutoring binary. The original git repository had the 2018-final-...-heroku/ folder as the root for the heroku backend. This should be sufficient to give you the gist of how it worked

NOTE: In ENV/lib I compressed all the python 3.6 stuff to keep under the 100 file limit. Should just need to be uncompressed